+++ src/config/data/ShortcutTable.hpp
@@ -38,6 +38,8 @@ public:
         Transpose,
         Reverse,
         ReplaceInstrument,
+        GrowPattern,
+        ShrinkPattern,
         Play,
         PlayFromStart,
         PlayFromCursor,